Restructure arch packages
