+++ app/src/main/java/com/futurice/freesound/arch/mvi/ViewModelExtensions.kt
@@ -30,7 +30,6 @@ fun <T> Flowable<T>.asUiModelFlowable(): Flowable<T> {
     return onBackpressureLatest()
 }
 
-
 // Events from the UI are modelled as a buffering Observable.
 fun <T> Observable<T>.asUiEventFlowable(): Flowable<T> {
     return toFlowable(BackpressureStrategy.BUFFER)
