+++ app/src/main/java/com/futurice/freesound/arch/mvvm/SimpleDataBinder.java
@@ -14,18 +14,24 @@
  * limitations under the License.
  */
 
-package com.futurice.freesound.arch.viewmodel;
+package com.futurice.freesound.arch.mvvm;
 
-interface LifecycleDataBinder extends DataBinder {
+import android.support.annotation.NonNull;
 
-    void onCreate();
+import io.reactivex.disposables.CompositeDisposable;
 
-    void onResume();
+/**
+ * Implements methods so that child classes don't need to declare empty implementations.
+ */
+public class SimpleDataBinder implements DataBinder {
 
-    void onPause();
+    @Override
+    public void bind(@NonNull final CompositeDisposable disposables) {
+        // Override this when required.
+    }
 
-    void onDestroyView();
-
-    void onDestroy();
-
+    @Override
+    public void unbind() {
+        // Override this when required.
+    }
 }
