+++ app/src/main/java/com/futurice/freesound/arch/mvvm/viewholder/AbstractBindingViewHolder.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.futurice.freesound.arch.viewmodel.viewholder;
+package com.futurice.freesound.arch.mvvm.viewholder;
 
-import com.futurice.freesound.arch.viewmodel.ViewModel;
+import com.futurice.freesound.arch.mvvm.ViewModel;
 
 import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
